Cast `.minimum` and `.maximum` to the same dtype as the `BoundedArray` spec
Allowing the dtypes of the bounds to differ from that of the spec instance can lead to potentially surprising behavior. For example, it is common practice to normalize arrays by subtracting the minimum and dividing by the range. If `spec.dtype` is float32 but `minimum` and `maximum` are float64 then the resulting array would be upcast to float64.

PiperOrigin-RevId: 260906207
Change-Id: Ie01ff31f961d276e2344afc4c0ad93b3c0df8c49