Add `specs.StringArray`, representing an `np.object` array of native Python strings
Using a plain `Array` to represent variable-length strings is problematic, since numpy's string dtypes can only have fixed lengths. An `np.object` array may be used as a container for native Python unicode or byte strings that can arbitrary lengths. However, by itself a dtype of `np.object` is ambiguous, since in principle such an array could contain any type of object.

To preserve the "string" semantics we introduce a special `StringArray` subclass of `Array`, which always exposes a dtype of `np.object`. This class also encapsulates custom logic for generating and validating values. `StringArray`s can contain either unicode or ASCII strings. The string type is specified in the constructor, and all elements in the array are required to have the same type.

PiperOrigin-RevId: 306682890
Change-Id: Ia9ddd96b5fcd296ddf4f949d5be1bd34c02d133d